Ai-chat-phase4/phase-4-ai-chat/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlmodel==0.0.22
pydantic==2.10.0
pydantic-settings==2.6.0
openai==1.58.0
pytest==8.3.0
pytest-asyncio==0.24.0
python-dotenv==1.0.1
PyJWT==2.10.0
aiosqlite==0.20.0